Cập nhật kiểu dữ liệu ID trong ThuocRepository và ThuocService từ Long sang Integer để đồng bộ hóa với thực thể Thuoc.

+++ src/main/java/com/ptithcm/sqa/repository/ThuocRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ptithcm.sqa.entity.Thuoc;
 
-public interface ThuocRepository extends JpaRepository<Thuoc, Long> {
+public interface ThuocRepository extends JpaRepository<Thuoc, Integer> {
     // Additional query methods (if needed)
 }
